+++ hex-literal/CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## UNRELEASED
+## 0.4.0 (2023-04-02)
 ### Changed
 - Disallow comments inside hex strings ([#816])
 - Migrate to 2021 edition and bump MSRV to 1.57 ([#816])